requirements.txt
aniso8601==8.0.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.29
protobuf==3.20.1
pycodestyle==2.8.0
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2020.1
requests==2.27.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
urllib3==1.26.9
Werkzeug==1.0.1
